# swagger.yaml
swagger: "2.0"
info:
  description: "This is the definitions for the Store Apis for the Invilla Challenge"
  version: "1.0.0"
  title: "Invilla Challenge - André Cintra (Store Api)"

paths:
  /store:
    post:
      tags:
      - "store"
      summary: "Create new store"
      description: ""
      operationId: "addStore"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "The store object that will be created"
        required: true
        schema:
          $ref: "#/definitions/store"
      responses:
        201:
          description: "Created"
  /store/{id}:
    get:
      tags:
      - "store"
      summary: "Retrieves store by id"
      description: "Returns a map with the store information"
      operationId: "getId"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of store that needs to be fetched"
        required: true
        type: "string"
        format: "uuid"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/store"
    patch:
      tags:
      - "store"
      summary: "patch the information for an store"
      description: "Path the information on the request body for the store with thi sid"
      operationId: "patchId"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Id of the store that needs to be patched"
        required: true
        type: "string"
        format: "uuid"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/store"

definitions:
  store:
    type: "object"
    properties:
      id:
        type: "string"
        format: "uuid"
      name:
        type: "string"
      address:
        type: "string"
  ApiResponse:
    type: "object"
    properties:
      id:
        type: "string"
        format: "uuid"
      name:
        type: "string"
      address:
        type: "string"